requirements.txt
django
django-mediaelementjs
djangorestframework
django-model-utils
django-environ
mysqlclient
gunicorn
jinja2
pytz
python-dateutil
django-jinja
openpyxl
requests
beautifulsoup4
lxml
zeep
pandas
bokeh
rdflib
django-debug-toolbar
celery[redis]
django-ckeditor-5
moviepy
pytest-django